requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
azure-core==1.36.0
azure-functions==1.24.0
azure-identity==1.25.1
azure-storage-blob==12.27.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
dotenv==0.9.9
fastapi==0.124.0
h11==0.16.0
idna==3.11
isodate==0.7.2
MarkupSafe==3.0.3
msal==1.34.0
msal-extensions==1.3.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uvicorn==0.38.0
Werkzeug==3.1.4